Factor OpenSSL11-specific code into its own file
